pipelines/requirements.txt
pandas
requests
mediacloud
lxml
psycopg2-binary
python-dotenv
newspaper3k
beautifulsoup4
rich
